requirements.txt
datasets==2.20.0
einops==0.7.0
ipykernel==6.26.0
ipywidgets==8.1.1
huggingface-hub==0.21.2
matplotlib==3.8.0
notebook==7.0.6
numpy==1.25.2
pandas==2.1.1
python-dotenv==1.0.0
scikit_learn==1.3.1
scipy==1.11.3
statsforecast==1.6.0
statsmodels==0.14.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.3
wandb==0.16.4